Move boolean simplification & remove rewrite.
